Moved ProjectForm to the main folder
