+++ Clients/src/presentation/components/Forms/ProjectForm/index.tsx
@@ -40,7 +40,7 @@ import DatePicker from "../../../components/Inputs/Datepicker";
 import dayjs, { Dayjs } from "dayjs";
 import { checkStringValidation } from "../../../../application/validations/stringValidation";
 import selectValidation from "../../../../application/validations/selectValidation";
-import CustomizableToast from "../../Toast";
+import CustomizableToast from "../../../vw-v2-components/Toast";
 import KeyboardArrowDown from "@mui/icons-material/KeyboardArrowDown";
 import { extractUserToken } from "../../../../application/tools/extractToken";
 import { useSelector } from "react-redux";
